FSV:2.0.1.0, V:2.0.1.0

+++ Clm/Clm/ChartData.fs
@@ -27,8 +27,8 @@ module ChartData =
         {
             modelDataId : ModelDataId
             binaryInfo : BinaryInfo
-            y0 : double
-            tEnd : double
+            y0 : decimal
+            tEnd : decimal
         }
